Multiple write caches
